requirements.txt
google-api-python-client
litellm
python-dotenv
pandas
rich
weasyprint
markdown
isodate
